requirements.txt
pyside6==6.7.1
pillow
numpy
pynput
google-genai
